SSCS-10116 Refactoring out methods from models
