--- src/main/java/uk/gov/hmcts/reform/sscs/model/Hashable.java
@@ -1,6 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model;
-
-public interface Hashable {
-
-    public Integer getHash();
-}
