+++ src/main/java/uk/gov/hmcts/reform/sscs/service/SessionCategoryMapService.java
@@ -1,12 +1,12 @@
 package uk.gov.hmcts.reform.sscs.service;
 
-import static uk.gov.hmcts.reform.sscs.helper.HashMapHelper.generateHashMap;
 import static uk.gov.hmcts.reform.sscs.helper.ReferenceDataHelper.getReferenceData;
-import static uk.gov.hmcts.reform.sscs.model.SessionCategoryMap.getHash;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.stereotype.Component;
@@ -17,15 +17,21 @@ import uk.gov.hmcts.reform.sscs.model.SessionCategoryMap;
 @Getter
 @Setter
 @Component
-public class SessionCategoryMapsComponent {
+public class SessionCategoryMapService {
     private static final String JSON_DATA_LOCATION = "reference-data/session-category-map.json";
+
+    private static final String SERVICE_CODE = "BBA3";
 
     private List<SessionCategoryMap> sessionCategoryMaps;
     private Map<Integer, SessionCategoryMap> hashMap;
 
-    public SessionCategoryMapsComponent() {
+    public SessionCategoryMapService() {
         sessionCategoryMaps = getReferenceData(JSON_DATA_LOCATION, new TypeReference<>() {});
         hashMap = generateHashMap(sessionCategoryMaps);
+    }
+
+    private Map<Integer, SessionCategoryMap> generateHashMap(List<SessionCategoryMap> sessionCategoryMaps) {
+        return sessionCategoryMaps.stream().collect(Collectors.toMap(this::getHash, reference -> reference, (a, b) -> b));
     }
 
     public SessionCategoryMap getSessionCategory(String benefitCode, String issueCode, boolean secondDoctor, boolean fqpmRequired) {
@@ -34,5 +40,21 @@ public class SessionCategoryMapsComponent {
 
     public SessionCategoryMap getSessionCategory(BenefitCode benefitCode, Issue issue, boolean secondDoctor, boolean fqpmRequired) {
         return hashMap.get(getHash(benefitCode, issue, secondDoctor, fqpmRequired));
+    }
+
+    public Integer getHash(SessionCategoryMap sessionCategoryMap) {
+        return getHash(sessionCategoryMap.getBenefitCode(), sessionCategoryMap.getIssue(), sessionCategoryMap.isSecondDoctor(), sessionCategoryMap.isFqpmRequired());
+    }
+
+    public Integer getHash(BenefitCode benefitCode, Issue issue, boolean secondDoctor, boolean fqpmRequired) {
+        return Objects.hash(benefitCode, issue, secondDoctor, fqpmRequired);
+    }
+
+    public String getCategoryTypeValue(SessionCategoryMap sessionCategoryMap) {
+        return String.format("%s-%03d", SERVICE_CODE, sessionCategoryMap.getBenefitCode().getCcdReference());
+    }
+
+    public String getCategorySubTypeValue(SessionCategoryMap sessionCategoryMap) {
+        return String.format("%s%s", getCategoryTypeValue(sessionCategoryMap), sessionCategoryMap.getIssue().name());
     }
 }
