+++ src/test/java/uk/gov/hmcts/reform/sscs/model/HearingDurationTest.java
@@ -2,6 +2,8 @@ package uk.gov.hmcts.reform.sscs.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.BenefitCode.UC;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.Issue.US;
 
 import java.util.List;
 import org.junit.Before;
@@ -9,15 +11,15 @@ import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
 import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitCode;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Issue;
-import uk.gov.hmcts.reform.sscs.service.HearingDurationsComponent;
+import uk.gov.hmcts.reform.sscs.service.HearingDurationsService;
 
 public class HearingDurationTest {
 
-    HearingDurationsComponent hearingDurations;
+    HearingDurationsService hearingDurations;
 
     @Before
     public void setup() {
-        hearingDurations = new HearingDurationsComponent();
+        hearingDurations = new HearingDurationsService();
     }
 
     @DisplayName("When valid Benefit Code and Issue Code is given to getHearingDuration the valid mapping is returned")
@@ -27,34 +29,16 @@ public class HearingDurationTest {
 
         assertEquals(BenefitCode.PIP_REASSESSMENT_CASE, result.getBenefitCode());
         assertThat(result.getIssue()).isEqualTo(Issue.LE);
-    }
-
-    @DisplayName("When an empty list of elements Disputed is given to getDurationFaceToFace"
-            + "the valid duration is returned")
-    @Test
-    public void getDurationFaceToFace() {
-        int result = hearingDurations.getHearingDuration("001", "US")
-                .getDurationFaceToFace(List.of());
-
-        assertThat(result).isEqualTo(30);
-    }
-
-    @DisplayName("When an empty list of elements Disputed is given to getDurationInterpreter"
-            + "the valid duration is returned")
-    @Test
-    public void getDurationInterpreter() {
-        int result = hearingDurations.getHearingDuration("001", "US")
-                .getDurationInterpreter(List.of());
-
-        assertThat(result).isEqualTo(60);
+        assertThat(result.getDurationFaceToFace()).isEqualTo(60);
+        assertThat(result.getDurationInterpreter()).isEqualTo(90);
+        assertThat(result.getDurationPaper()).isEqualTo(30);
     }
 
     @DisplayName("When an empty list of elements Disputed is given to getDurationFaceToFace"
             + "the valid duration is returned")
     @Test
     public void addExtraTimeNoElement() {
-        int result = hearingDurations.getHearingDuration("001", "US")
-            .addExtraTimeIfNeeded(30, List.of());
+        int result = hearingDurations.addExtraTimeIfNeeded(30, UC, US,List.of());
 
         assertThat(result).isEqualTo(30);
     }
@@ -63,8 +47,8 @@ public class HearingDurationTest {
             + "the valid duration is returned")
     @Test
     public void addExtraTimeCorrectElement() {
-        int result = hearingDurations.getHearingDuration("001", "US")
-                .addExtraTimeIfNeeded(30,List.of("RY"));
+        int result = hearingDurations.addExtraTimeIfNeeded(30, UC, US,List.of("RY"));
+
         assertThat(result).isEqualTo(30);
     }
 
@@ -72,8 +56,7 @@ public class HearingDurationTest {
             + "the valid duration with extra time is returned")
     @Test
     public void addExtraTimeCorrectWc() {
-        int result = hearingDurations.getHearingDuration("001", "US")
-                .addExtraTimeIfNeeded(30,List.of("WC"));
+        int result = hearingDurations.addExtraTimeIfNeeded(30, UC, US,List.of("WC"));
 
         assertThat(result).isEqualTo(45);
     }
@@ -82,8 +65,7 @@ public class HearingDurationTest {
             + "the valid duration with extra time is returned")
     @Test
     public void addExtraTimeCorrectSg() {
-        int result = hearingDurations.getHearingDuration("001", "US")
-                .addExtraTimeIfNeeded(30,List.of("SG"));
+        int result = hearingDurations.addExtraTimeIfNeeded(30, UC, US,List.of("SG"));
 
         assertThat(result).isEqualTo(45);
     }
